refactor(diyui): Panel wrapper 类名去 Panel 前缀，PanelApp 独立文件，命名空间对齐 Panel 原生
