+++ packages/diyui.py/tests/browser_test_app.py
@@ -4,9 +4,9 @@ run: uv run panel serve tests/browser_test_app.py --port 0
 """
 
 import diyui
-import diyui.providers.panel
+import diyui.providers.panel as diypn
 
-app = diyui.providers.panel.PanelApp(config=diyui.ScopeConfig(mode="dev", scheduler=diyui.ImmediateScheduler()))
+app = diypn.PanelApp(config=diyui.ScopeConfig(mode="dev", scheduler=diyui.ImmediateScheduler()))
 
 with app.column():
     app.markdown("# 🧪 Browser Test App")
@@ -20,7 +20,7 @@ with app.column():
     )
 
     @app.column().cell()
-    def _(node: diyui.providers.panel.PanelColumn):
+    def _(node: diypn.layout.Column):
         app.markdown(f"## Hello **{name_input.value}** × {multiplier_input.value} !")
         app.markdown(f"Repeated: {'🔥' * multiplier_input.value}")
 
@@ -33,7 +33,7 @@ with app.column():
         btn_inc.on_click(lambda e: setattr(counter, "value", counter.value + 1))
 
     @app.column().cell()
-    def _(node: diyui.providers.panel.PanelColumn):
+    def _(node: diypn.layout.Column):
         app.markdown(f"Counter: **{counter.value}**")
